Plugin structure
Move scripts that process the scale data into the plugins folder and
make sure that BS440.py doesn't need to be changed upon the arrival of
new plugins. First plugin: BS440mail [untested]
