+++ BS440.ini
@@ -3,20 +3,6 @@
 # Change to your preference
 # Store this file in the working folder of BS440.py
 
-# The persons secion lists all the scale users
-# The list maybe extended to 8 persons
-
-[Person1]
-username: John
-useremail: johndoe@gmail.com
-googleauthfile: google-john.json
-csvfile: john.csv
-
-[Person2]
-username: Jill
-useremail: jilldoe@gmail.com
-googleauthfile: google-jill.json
-csvfile: jill.csv
 
 # Scale properties
 [Scale] 
@@ -29,23 +15,3 @@ device_name: 0202B6332211CCBBAA
 loglevel: debug
 logfile: BS440.log
 
-# Sender email/pwd: gmail account info
-# Uncomment [Email] to have emails sent
-[Email]
-sender_email: yourmail@gmail.com
-sender_pwd: yourpassword
-
-# Uncomment [Domoticz] to enable updating
-# DomoticzIP: IP adres of Domoticz server
-# Assumes unauthenticated access when on local LAN
-# set in Domoticz|setup|Settings|System|Local Networks
-# hardware_name is optional (default: Medisana)
-#[Domoticz]
-#domoticz_url: 127.0.0.1:8080
-#hardware_name: 
-
-#Uncomment [Google] to enable Google FIT updating
-#[Google]
-
-# Uncomment [CSV] to use CSV
-#[CSV]
