+++ plugins/BS440mail.ini
@@ -1,0 +1,19 @@
+# BS440mail.ini
+# Private Initialization file for BS440mail.py
+# Change to your preference
+# Store this file in the plugins folder with BS440mail.py
+
+# The persons section lists all the scale users
+# The list maybe extended to 8 persons
+
+[Person1]
+username: John
+useremail: johndoe@gmail.com
+
+[Person2]
+username: Jill
+useremail: jilldoe@gmail.com
+
+[Email]
+sender_email: yourmail@gmail.com
+sender_pwd: yourpassword
